+++ src/hooks/useGenericSearch.ts
@@ -4,9 +4,9 @@ import { UseExtendQueryOptions, useQuery } from '@/utils/api';
 
 import { GenericResource } from '../types';
 
-import { getQueryKey } from './queryKeys';
 import { mapAssetStataus } from './useAssets';
 import { useAxios } from './useAxios';
+import { getQueryKey } from './useQueryKeys';
 
 export const useGenericSearch = (
   {
